feat: v3.5.22 — IPC abstraction layer, Windows daemon support, CPU hotfix #210
- New ipc/ module (mod.rs, process.rs, unix.rs, windows.rs) isolates all
  platform-specific daemon code. Unix UDS + Windows Named Pipes behind
  DaemonAddr enum. No more #[cfg(unix)] in daemon.rs/daemon_client.rs.
- HTTP-based daemon shutdown (POST /v1/shutdown) replaces OS signals.
- build_app_router() extraction eliminates serve()/serve_uds() duplication.
- Hotfix #210: env.sh recursion guard + graph scan limits (50k entries,
  2min timeout, cross-process lock, is_safe_scan_root).
- Parallel call graph build with rayon + live progress tracking.
- Dashboard: call graph progress bar, project file browser in Compression
  Lab with search, graph-files API endpoint.

+++ rust/src/ipc/process.rs
@@ -26,7 +26,7 @@ pub fn is_alive(pid: u32) -> bool {
             let mut exit_code: u32 = 0;
             GetExitCodeProcess(handle, &mut exit_code);
             CloseHandle(handle);
-            exit_code == STILL_ACTIVE
+            exit_code == STILL_ACTIVE as u32
         }
     }
 }
